[ToolsCommon] Break out ReturnCode
Before, if any file wanted to use ReturnCode, it's project would need
to include all dependencies of Utils.cs, regardless of usage. By
moving ReturnCode into it's own source, we break the dependency.

+++ src/Tools/dotnet-trace/CommandLine/Commands/CollectCommand.cs
@@ -472,7 +472,7 @@ namespace Microsoft.Diagnostics.Tools.Trace
             {
                 Console.Error.WriteLine($"[ERROR] {dte.Message}");
                 collectionStopped = true;
-                ret = dte.ReturnCode;
+                ret = (int)dte.ReturnCode;
             }
             catch (OperationCanceledException)
             {
